Working on delete, changing show

+++ src/main/java/Book.java
@@ -1,5 +1,5 @@
 public class Book extends Media {
-    Book(String name, int condition, String description, String media, int price) {
+    Book(String name, int condition, String description, String media, double price) {
         super(name, condition, description, media, price);
     }
 }
