+++ src/main/java/Game.java
@@ -1,5 +1,5 @@
 public class Game extends Media {
-    Game(String name, int condition, String description, String media, int price) {
+    Game(String name, int condition, String description, String media, double price) {
         super(name, condition, description, media, price);
     }
 }
